+++ apps/console/src/features/identity-providers/configs/ui.ts
@@ -1,10 +1,19 @@
 /**
- * Copyright (c) 2019-2023, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ * Copyright (c) 2019-2023, WSO2 LLC. (https://www.wso2.com).
  *
- * This software is the property of WSO2 LLC. and its suppliers, if any.
- * Dissemination of any information or reproduction of any material contained
- * herein in any form is strictly forbidden, unless permitted by WSO2 expressly.
- * You may not alter or remove any copyright or other notice from copies of this content.
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 
 import { identityProviderConfig } from "../../../extensions/configs/identity-provider";
@@ -77,10 +86,19 @@ export const getIdentityProviderWizardStepIcons = (): any => {
 export const getAuthenticatorIcons = (): any => {
 
     return {
+        AppleOIDCAuthenticator: AppleLogo,
+        BasicAuthenticator: BasicAuthIcon,
+        FIDOAuthenticator: FIDOLogo,
+        FacebookAuthenticator: FacebookLogo,
+        GithubAuthenticator: GithubIdPIcon,
+        GoogleOIDCAuthenticator: GoogleLogo,
+        MagicLinkAuthenticator: MagicLinkLogo,
+        Microsoft: MicrosoftLogo,
         apple: AppleLogo,
         backupCode: BackupCodesAuthenticatorLogo,
         basic: BasicAuthIcon,
         default: ConnectionIcon,
+        "email-otp-authenticator": EmailOTPIcon,
         emailOTP: EmailOTPIcon,
         facebook: FacebookLogo,
         fido: FIDOLogo,
@@ -91,6 +109,7 @@ export const getAuthenticatorIcons = (): any => {
         microsoft: MicrosoftLogo,
         office365: Office365Logo,
         sessionExecutor: ClockColoredIcon,
+        "sms-otp-authenticator": SMSOTPIcon,
         smsOTP: SMSOTPIcon,
         totp: TOTPIcon,
         twitter: TwitterLogo,