+++ apps/console/src/features/org-insights/config/org-insights.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ * Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -61,6 +61,18 @@ export const getFilterAttributeListByActivityType = (activityType: ActivityType)
                     "console:manage.features.insights.activityType.login.filters.userStore"
                 ) as ReactNode,
                 value: "userstoreDomain"
+            },
+            {
+                text: I18n.instance.t(
+                    "console:manage.features.insights.activityType.login.filters.authenticator.attributeName"
+                ) as ReactNode,
+                value: "authenticator"
+            },
+            {
+                text: I18n.instance.t(
+                    "console:manage.features.insights.activityType.login.filters.identityProvider"
+                ) as ReactNode,
+                value: "identityProvider"
             }
         ],
         [ActivityType.REGISTRATION]: [