+++ apps/myaccount/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @wso2is/myaccount
 
+## 2.0.6
+
+### Patch Changes
+
+- [#4135](https://github.com/wso2/identity-apps/pull/4135) [`9bb63bf50e`](https://github.com/wso2/identity-apps/commit/9bb63bf50e805d423d8094cb79a9b0d125041255) Thanks [@JayaShakthi97](https://github.com/JayaShakthi97)! - Improvements and bug fixes
+
+- Updated dependencies [[`9bb63bf50e`](https://github.com/wso2/identity-apps/commit/9bb63bf50e805d423d8094cb79a9b0d125041255)]:
+  - @wso2is/theme@2.0.4
+  - @wso2is/core@2.0.2
+  - @wso2is/i18n@2.0.2
+  - @wso2is/react-components@2.0.3
+
 ## 2.0.5
 
 ### Patch Changes