+++ modules/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @wso2is/core
 
+## 2.0.2
+
+### Patch Changes
+
+- [#4135](https://github.com/wso2/identity-apps/pull/4135) [`9bb63bf50e`](https://github.com/wso2/identity-apps/commit/9bb63bf50e805d423d8094cb79a9b0d125041255) Thanks [@JayaShakthi97](https://github.com/JayaShakthi97)! - Improvements and bug fixes
+
 ## 2.0.1
 
 ### Patch Changes